Simplify UNIX configuration
Compile-testing on FreeBSD 8.2 showed that it didn't matter wether
BSD or SYSV was defined, but defining LINUX was required. Setting SVR4
didn't seem to make a difference.
On Linux, SYSV && LINUX was required.

So, BSD and it's code has been removed; SYSV, LINUX and SVR4 are
unconditional and the defines are gone