initial separation of libnethack from the tty window code
headers have been split into public and private headers,
with the public headers now in include/public/
A new header nethack.h defines all the publicly exported
functions and variables, while export.h and decl.y are for
internal stuff.
libnethack now builds with -fvisibility=hidden, to prevent the use
of functions that have not been intentionally exported

The result is not really pretty, but at least the dependencies are now
visible. More effort is needed to make this a real API.