rewrite menu interaction
the functions start_menu, add_menu, end_menu and select menu have been
replaced by display_menu and display_objects which operate on lists of
items.
This also made it possible to get rid of the "anything" union.

All of the above are still present in the tty port, but only as internal
implementation details.