new savegame format with move logging
each save now contains a running log of all actions in the game which
can be replayed to perfectly recreate the last game state. If the game
is saved normally, this log is followed by a regular binary save.

The log makes the INSURANCE code and the recover utility obsolete and
can also replace ttyrecs - except that it will also work for tiles mode.

The binary save is there to make loading long-running games fast: the
log can be replayed at about 20k moves/second (on my machine), which
would still cause some delay when restoring an ascension run.