Use a separate rng for display randomization while Hallucinating
Unfortunately the code paths for restoring vs replaying differ
significantly where displaying is concerned because the display buffer
and the vision state are not saved.
Normally this does not matter, because recomputing them does not affect
the game state. This changes when the player is hallucinating. Restoring
such a game will cause a large number of "extra" rng calls, which would
otherwise break the game.