nh_get_object_commands: list suitable commands to apply to an object
This code was derived from the itemactions code in AceHack and
UnNetHack.
The difference is that the list is not displayed directly to the player
but returned to the calling function. Since it is part of the API, it
allows the UI code to create item action menus or right click menus
or whatever makes sense for the specific UI.