autopickup rules: the replacement for pickup_types and AUTOPICKUP_EXCEPTIONS
The autopickup rules are only active if autopickup is on. If autopickup
is on and you walk onto an item, the item is compared to each rule in
turn until a matching rule is found. If a match is found, the action
specified in that rule is taken and no other rules are considered.
If no rule matches, the item is not picked up.

Each rule may match any combination of object name, object type,
and object blessing (including unknown).
A rule that specifies none of these matches everything.

Example rules as shown in the UI:
 1. IF name matches "*lizard*" AND type is "food": < GRAB
 2. IF name matches "*corpse*" AND type is "food":   LEAVE >
 3. IF type is "food":                             < GRAB

A newt corpse, for example, will not match rule 1, but rule 2 applies,
so it won't be picked up and rule 3 won't be considered.