Fix findtravelpath buffer overflow bug (from AceHack)
The easiest way to recreate the bug is on the UNIX tty version of
vanilla NetHack; in wizmode, dig out a level (or force the bigroom
variant that's already an entirely dug-out level to generate), then fill
the entire thing with boulders (it takes a while). Then try to travel
from one corner to the diagonally opposite corner, and you'll get a
segfault.

The bug's in the conditional starting at hack.c:700. The enclosing for
loop is run once for each direction (8 times total), and sets nx and ny
for the next x and y in the direction it's considering. The conditional
on line 700, though, checks Passes_walls (independent of dir/nx/ny),
can_ooze (likewise independent), closed_door(x, y) (likewise
independent), and sobj_at(BOULDER, x, y) (likewise independent); in
short, whether it's true or not has nothing to do with dir, nx, or ny,
and thus it'll either be true on all 8 iterations, or false on all 8
iterations.

The case where it's false is the usual case, and the code in that case
works correctly. If it's true, though, it counts the square as costing 3
rather than 1, ("/* closed doors and boulders usually cause a delay, so
prefer another path */"), and the way it does this is to re-add the
current entry in travelstepx/y to the next step's current entry in
travelstepx/y, up to twice more. Although this has the desired effect of
delaying considering the step into the future, the step ends up getting
added eight times, once for each direction, even though it has nothing
to do with the direction. So instead of adding 1 step to the buffer, it
adds 1, then 8, then 64, and evaluates the next move 64 times from
there. (However, potential moves /from/ the square are only added once,
with the code seeing 1 optimal route and 63 suboptimal routes.)

The buffer, in theory, has enough room to contain every square on the
map once. It /doesn't/ have enough room to contain every square on the
map 64 times; but this typically doesn't matter, as there's room for
COLNO*ROWNO (80*21 or 1680) steps in the buffer, which holds all the
steps at the same distance from the player. (It should be ordinary
squares counting 1, boulders/doors counting 1; but due to the bug,
ordinary squares count 1, and boulders/doors count 64.) The easiest way
to trigger the bug, therefore, is for 27 or more boulders/closed doors
to be equidistant from the player, leading to 1728 squares being stored
in the length-1680 travelstepx array.

In practice, though, considerably more boulders are needed for the
overflow to have any visible effects, because findtravelpath allocates
several large arrays on the stack, and they tend to give some overflow
room. That's why doing something as dramatic as filling the level with
boulders is needed in order to produce an obvious demonstration of the
bug.