separate starting role from the role set in the options
This fixes starting a second game in a process where a previous game
already ran.
Initially, flags.init* are set according to the options.
Then, during startup of a game, these flags were set to the actual
values chosen by the player, so none of them were ROLE_NONE or
ROLE_RANDOM any more.
Then, when starting a second game, the player would no longer get to
choose their character, because the race, role, gender & alignment were
already fully specified.

This change separates the flags into flags.init* for the values from the
options and u.init* for the values used to start the running game.