NitroHack Network Server - network games without telnet!
The server will support netwok games with classical text-mode clients,
but also with tiles clients and even web browsers.
In order to do this, the libnitrohack API calls are translated to JSON
and sent over the Network between client and server.

The client network library is not complete yet, so the server barely
tested and not yet usable for actual games.

Requirements: Linux, Jansson (JSON library), PostgreSQL client library.