limit the number of times the client will attempt to resend a command
A player found a crash bug on the server. His client detected the crash
as a connection loss and reconnected, which caused the server to fork a
new game process. Then the client resent the command and the game
process crashed again.
With unlimited retries, this effectively resulted in a hang.