make it possible to view running games (but not "live")
Running games can now be selected on the view menu, but they behave as
if they had been saved: they can be viewed up to the point that was
reached when the game was loaded for viewing, but newer moves don't become
available for viewing automatically.
Unfortunately, the strict protocol checking in 4.0.0 makes it impossible
for the protocol to remain compatible. Therefore the default port has
been changed to 7115 (from 7114) so that old and new server can co-exist
on the same host.