add a newline after the end of the game log when the log is closed
Otherwise the topten status at the end of the file shares a line with
the final command status.

As a side effect, this change makes it possible to see  when the game
was saved: the log will have an extra newline at that point.
Maybe that will make reproducing complicated save problems easier.