fix: replay & restore might break due to unspecified behavior
Specifically the order in which the arguments of a function call are
evaluated is not specified and can vary between compilers and
optimization levels.
Therefore all instances of some_func(random(x), random(y)) can break
replay when using a different binary.
This change makes the ordering of random calls explicit in the level
generation code; any save relying on a different ordering is now broken.