Display menu/question/prompt selections in message history
In the message area, too, if there's room (there isn't on 80x24, but
there is if we start auto-moring). This requires a breaking change to
the window-port API, to display nonblocking messages (which is an
AceHack feature, originally introduced for farlook, but which turned
out to be useful for other things too).