Track player's knowledge of locked/unlocked doors, draw it onscreen
This includes a breaking API change, to allow square brandings to be
sent over the windowport API. Currently it sends locked/unlocked,
trapped/untrapped, and stepped; but only locked/unlocked/trapped are
drawn, and only on closed doors. (AceHack fans may know what this is
leading up to...)

Based on a change from AceHack.