Limit the high score display to BUFSZ columns.
This is plenty wide enough and has the added benefit of preventing a buffer
overflow when the menu items are larger than BUFSZ wide.

I would really like to put an assertion into set_menuitem to avoid this
happening in the future, but there doesn't appear to be any mechanism for
something like this?