Fix C343-34.
Mimcs now mimic open doorways on the rogue level if they somehow get there. It
doesn't make a whole lot of sense in particular, but it's better than other
options, I think. Mimcs aren't generated on the rogue level anyway, so it will
take some work for this to be noticeable.