Fix a bunch of compiler warnings.
Of note, I've changed the one- and sometimes zero-length array members we used
for extending structs to have undefined length. This is a C99 feature. I don't
know if MSVC can handle it; if not, zero is fine. I've also changed some of them
to pointers, since pointers will be correctly 8-bit aligned, which matters on a
64-bit system as it keeps the pointers in the structures stored in the padding
space from being incorrectly aligned.