Support automatic merging in inventory via #adjust
In vanilla nethack, adjusting an object to a slot containing an
object it can merge (including specifying the same inventory letter
twice) will merge all objects in inventory into a single stack.
This is handy in cases like when you've just diluted a large number
of potions, and want to stack them all.

Although the new implementation is probably a bit more versatile,
it lost that feature.  This patch adds a case for if the same
inventory letter is specified twice (which right now does nothing
useful) it merges everything it can into that stack.

One thing I noticed while testing that was that objects of the
same BUC status will merge, even if the BUC status of some is
known and for others is not.  From my reading of the code, it looks
like that change was probably intentional, so I didn't change it,
but it doesn't seem right to me.