Deliver objects falling through a trapdoor to the correct level.
Also fix messages for boulders and statues destroyed out of sight.

Some of the functions used by deliver_object assumed that the object
was being delivered to the current level.  This fixes the cases that
are used by deliver_object to use the object's level instead.  Also,
messages for boulders and statues destroyed out of the player's sight
were being issued even if it wasn't visible; fixed that here because
it touches some of the same code.