The Mass Reindent
We now have a defined indentation style (explained in indent(1)
format in indent.pro, and approximated in an Emacs and Vim modeline
in every .c and .h file). All C and header files were automatically
updated to use it, and some of the worse mistakes made by the
autoindentation program fixed by hand. (There are probably several
more lurking; if anyone sees a bizarre bit of indentation, they
should feel free to change it to something more sensible.)

The style itself is basically four-space indents with 1TBS; tabs
shouldn't appear in .c/.h files, but if one does, it means eight
spaces (two indents), and needs to be converted to spaces.