Fix shopkeeper msg on death when not seen
This fixes a logic error where a misplaced parenthesis sends a boolean
to cansee() where a y-coordinate is expected, causing the wrong
shopkeeper to be noted in the death message in a store when there are
multiple shopkeepers on the level.

(cherry picked from commit 75150469502cee25b7ecc4014a2614c902303245)

Originally from UnNetHack CS530.

Conflicts:
	libnitrohack/src/shk.c