Loosen 'i' command dropping restrictions
This allows a few items that weren't previously allowed to be dropped to be
dropped, in particular a heavy iron ball with which you are punished. I think
this also affects alternate weapons.

For convenience, also added a W_WORN macro which represents an item you are
wearing anywhere on your body (but not one being wielded).