Improve monster !/oI message when telepathic.
Now gives a separate monster when a monster drinks a potion of invisibility or
zaps itself with a wand of make invisible if you cannot see the now-invisible
monster, but you can detect it telepathically. Previously, the message was
confusing because you were told that the monster disappears, but no change
occurred to the glyph.

No message change for monsters detected via skilled/expert monster detection
since the background colour of the monster will change, and the message doesn't
seem that odd.