Unset quiver on cancelled empty quiver fire prompt
This unsets the quiver if the player accidentally chooses an object to
quiver when firing from an empty quiver but cancels at the throw
direction prompt.  This prevents vi-keys players from having to manually
unset their quiver if they typo an object at that quiver prompt when
expecting a firing direction, which is otherwise quite annoying.

This is meant to go with commit 1115668bbc (prompt to quiver if firing
with an empty quiver).

(cherry picked from commit c4894124d6c206052554f00c3bc71827f01df906)

Conflicts:
	libnitrohack/src/dothrow.c