Add pay to inventory object action menu
An unpaid object in the inventory can be paid for from the new 'pay'
option in its inventory object actions menu.

Compared to UnNetHack, the code is completely original to account for
NitroHack's new command system.  The price mentioned for unpaid
containers with unpaid contents reflects only the container, since so
far that's the only object paid for when this command is used.

(cherry picked from commit 957e132df252353e538f0af71dcdbd0bdd537814)

Originally based on UnNetHack CS839.

Conflicts:
	libnitrohack/src/cmd.c
	libnitrohack/src/shk.c