Suggest #wipe when blinded by venom or cream pie
Like the #monster and #sit suggestions when polymorphed, this suggestion
only appears if the 'verbose' option is enabled.  Given how rare it is
for this command is useful, and how useful the command is when it is
useful, it's strange that NetHack doesn't suggest it already.

(cherry picked from commit 39e2c137f1cb25198dbe52955bb791a76a369e86)

Conflicts:
	libnitrohack/src/dothrow.c
	libnitrohack/src/mthrowu.c