Show weight in inventory object action menu title
If the invweight option is enabled, include the weight of objects in the
title of their action menu, like it appears in the inventory proper.

This also fixes a crash bug in that menu for objects with very long
names (the buffer was too large for the menu system to handle).

(cherry picked from commit 8f7e2fced9327e885c3570722df4c2e6a0dac2c7)

Conflicts:
	nitrohack/src/menu.c