Fix travel moving player back and forth repeatedly
This fixes a bug where hundreds of turns are wasted by the travel system
moving the player back and forth when the player targets an unreachable
space and sight-blocking obstacles occur in certain formations
in-between.  See the comment in the code for full details.

The use of couldsee() in conjunction with the smarter travel pathing
when guessing an intermediate travel target point was the cause of the
bug.  Other solutions considered:

 * Why not scrap the use of couldsee() altogether and just use the
   player-reachable space closest to the target?  The problem is that
   this could cause the player to be led halfway around the level to get
   to a space that's one closer than one very near the player's starting
   position.

 * Why not factor in how far the player moves in the above method?  The
   position that the game leads the player becomes very unpredictable.

The solution used limits the intermediate target while guessing to
spaces in the couldsee() 'blob' occupied by the player, cutting off
isolated couldsee() spaces to prevent them from being chosen, which
would otherwise cause the smarter travel system to move one way in one
turn while the couldsee() heuristic would move it back the other way on
the other turn.

This bug also exists in vanilla NetHack, where it can loop until the
player is fainting from hunger (this branch has a feature that limits
the bug to when the player reaches hungry).  As such it also exists in
UnNetHack and upstream NitroHack.  The specific fix is adapted for the
autoexplore changes in this branch, but can be easily retro-fitted into
vanilla NetHack.

(cherry picked from commit 02da53ea5ff738533e647e0ef4514e95b38206b0)

[AIS: I've been wondering what caused that bug for /ages/. Now we
know. Thanks for finding it!]

Conflicts:
	libnitrohack/src/hack.c