Autoexplore: Fix re-exploring items w/ ball/chain
Without this fix, dragging a ball or chain over an object pile would
cause autoexplore to re-explore it.  This would also happen if the
player were standing on the pile with the ball/chain, and when pets
picked up and dropped items.

The logic to re-explore an object pile if memory of it changes was too
naive.  Though the previous cases could be catered to, it's far more
simple and robust to simply not do that, even if the results aren't
perfect.

This is meant to go with commit 2f85168184 (autoexplore: don't stop for
explored object piles) and commit dffe879cc3 (autoexplore: fix repeated
exploring of ball+chain).

Fixes #52 on [DynaHack's tracker].

(cherry picked from commit 6bf8e3dd320402ae12585593e22c4dc5f6f35248)

Conflicts:
	libnitrohack/src/display.c