Refactor temporary map symbols.
Previously, the interface to temporary map symbols was a single tmp_at()
function that stored all its state internally. Initialization and reentrancy
were handled with a series of special command coordinates. This interface was
confusing and did not lend itself well to the impetus for this fix (see below).
Additionally, its handling of reentrancy was stack-based; while it could pause
one set of temporary symbols to handle another, it could not do them
interchangeably.

The new interface includes a storage struct held by the caller, but
memory-managed by the interface. This has several advantages: it is clearer, it
has better reentrancy properties (it can handle concurrent sets of temporary
symbols, although not if they cross paths), and it allows for fixing a bug.

Bug #3, the impetus for this change, arose due to the fact that corpsenm data
was not preserved through the tmp_at() interface, and this produced a buffer
underrun in the display engine. This now fixes this bug and thrown corpses
display correctly, with the right color.