Implement new wizard teleportation.
This changes the behaviour of wizard mode a bit by not giving guaranteed
teleport control and instead causing wizard-mode teleport commands to override
normal checks (like wizard mode always did previously).

In order to accomodate this, a new command ^R is added to invoke wizmode
teleport; ^T remains regular teleport.

The biggest upshot of this is you can test regular teleportation in wizmode now.

(If quantum mechanics are being annoying, wish for a blessed ring of tele
control, or tengu corpses, or the MKoT, or something. Really, there are enough
options available!)