Fix the infamous blue message bug, ticket #87.
I'm not entirely sure I agree with the message engine's distinction between
turns and actions, because of the fact that this means that high-multi actions
such as running or autoexploring get blue messages for things encountered along
the way. It would probably be better to cause the color of message to twig off
of the most recent input from the user, as this is far more likely to be
correct.

However, this is rather clearly the design of the code as written, and so I'm
going to say that the blue message bug itself has been fixed, and changing the
behaviour for multi actions should be the subject of a separate bug or
enhancement request.