Fix aimake build failure on Perl 5.18
I don't know why this fix works, and it worries me. (Or at least, I
know why it works in that the config file is trying to use rv2gv,
which wasn't whitelisted for config files before. I don't know why
it was trying to use that, though. Or, if it's for a good reason,
why it wasn't on previous versions of Perl.)