Fix full screen redraws
In particular, they weren't redrawing spaces, rather spouting
character set codes into the screen. This bug had no symptoms on many
terminals but a bit of slowness on startup, but for terminals that
don't parse "enter Unicode", it would fill all characters up to the
first non-space character with Gs.