Remove "colorless" as a color from the game
It was only being used for two things: being listed as a color name
when a random color name was wanted (and thus never needing to be
drawn), and for invisible monsters. The random color is now listed as
"dark gray" if that's rolled when a random color is wanted (again,
it's OK to name it even if the player has darkgray turned off);
invisible monsters are now shown as a light blue 'I', not a colorless
'I'.

Because colorless is not a real color, the client was rendering it as
dark gray (which had a numerically equal value of 8) even with
darkgray turned off. That caused bug #168, which should be fixed by
this change.