Update rules, dependencies, OS detection
We don't need a separate copy of eos() in dlb_main, because the one in
hacklib works fine (and if we have two copies, aimake 2 and 3 are confused
about which to use). The files with OS dependencies need adjusting to be able
to deal with them even without headers included, to avoid the chicken and egg
problem where the OS is needed to know which headers are included and the
included headers are needed to detect the OS. Also get the level compiler
hooked up to the build system.