Remove unnecessary dependencies
aimake is upgraded to a version that complains on more bad dependency
patterns, and many such patterns are fixed as a result. Probably the
largest change is to alter hacklib.c to actually match its
documentation (previously, it had too many dependencies on program
state); this is accomplished via splitting part of it out into a new
file localtime.c (and moving one function to windows.c). hacklib.[ch]
are now entirely self-contained, and thus can be linked by both src
and util without pulling in a huge number of unexpected dependencies.