Timestamps for license compliance
So far, we've only been distributing via VCS, which mixes the
authorship information for the files into the same packs that
everything else is distributed with. However, some day we might
distribute the files individually, in which case (according to
the NHGPL) they each need their own, individual, change details.

This change adds a last-modified author and date to each file in
the distribution, as well as a script that allows them to be
easily updated. (The intended workflow is to run the script and
commit its results just before each push.)

All the last-modified author/dates are me and today, because I
just changed the files to put the last-modified dates on them...