Give landmines and bear traps unidentified appearances
This works the same way as, say, elven daggers; there's only one
possibility for what a "toothed trap" in your inventory could be,
but it gets an amusing, descriptive flavour name for characters who
don't have experience with them (at the moment, setting or disarming
one, both actions informing the character of the relation between the
item and the trap).

This makes scoring marginally more interesting (now you have to get
hold of the traps for full item identification), but more
importantly, it gives a land mine (the item) and a land mine (the
trap) different unidentified names. This actually caused a minor but
very long-standing bug in NitroHack (that has existed in all versions
of NetHack 4); it would be impossible to customize the symbol or
colour of either sort of land mine, without affecting the other one
as well. The bug becomes more major with the implementation of tiles,
which is why it's being fixed now (and why I noticed it in the first
place).