Use unidentified item descriptions as their API names
Previously, to customize what a runed wand looked like, you'd have to
edit the config file entry for "wand of sleep", because wands of
sleep and runed wands are in the same slots on the IDed and unIDed
item description tables respectively. This was pretty confusing, and
made all the more so via the need to customize, say, "wand of unnamed
2" if you wanted to change the appearance of a spiked wand.

This patch changes item API names to work more like a user would
expect; item appearances are customized via their unidentified names
(e.g. "runed wand"). In the case of multiple items that have the same
textual appearance (e.g. "conical hat", "bag"), the appearances of
the individual items in the set are given names like "bags 0",
"bags 1", etc.. (At present, the code appears to avoid spoilers via
using a different appearance for IDed and unIDed items in the set,
which is a bit bizarre in practice. That was (unchanged) existing
behaviour; it might make more sense to shuffle appearances instead,
but that's a change for the future.)