Make the code that tracks what can exist in the game self-contained
Things like the list of all monsters in existence, their API names,
and so on, are useful to the tiles manipulation code as well as to
the game and the server. drawing.c and its dependencies are now
self-contained, and do not rely on any object files from libnethack
as a whole.