Use author not commiter timestamps in update-timestamps.sh
It's unclear what the letter of the license means with respect to a
VCS, anyway (which file, specifically, is being changed?); and using
the author timestamp seems to fit the spirit of the license better
(i.e. the time at which the file itself was changed, as opposed to
the time at which the change was added to the repository).