Working key input for SDL
We can't base our input on the key input (it doesn't have enough
information to resolve ambiguous keystrokes like shift-3, because it
operates on the level of keys on a keyboard); we also can't base it
on the text input API, because that doesn't report things like arrow
keys at all. In order to keep things flexible to even unknown input
methods, this uses a combination of the methods; if a key is pressed,
and SDL doesn't say it belongs to an input method within 2ms, nor
give a Unicode equivalent, we just use the key directly, otherwise we
wait until the Unicode for the key turns up.