Disable non-working font sizes and tile sizes
Switching fonts isn't currently implemented (and the obvious
implementation drives SDL crazy, so there's probably something wrong
in sdl.c that should be fixed later). 3D tiles aren't really
implemented at all (the tiles utilites understand them, but the
windowports don't); and both the 3D and 32x32 tiles are missing
images. (A necessary step before they can be used is to convert the
16x16 images to do transparency properly.)

Thus, for now, we don't install the non-working fonts/tilesets, and
don't give the user the option to select them.