Mass Reindent: stage 3, DynaHack edition
DynaHack uses different indentation (it's inconsistent, like vanilla
is); thus, we need to reindent the patches copied from it.

libuncursed and tilesets are omitted; neither exists in DynaHack, and
both are substantially changed in the windowports branch, making that
a better place to reindent them.

indent(1) was used to suggest indentations, which were then accepted
or rejected by hand. (In a few cases, the indentation was adjusted
manually.)