Make eating off the ground actually work
This fixes 03bd7a6dedcf2d72fca8eaf332dadf235053ba91,
"Choose floor items from a menu when eating, sacrificing and tinning".

The floorfood function wasn't returning the selected item; it was
just ignoring it and asking again.  This change also makes it not
prompt if there's only one item on the square.

This bug also exists in DynaHack, as far as I can tell.