Use recalculation, not dead-reckoning, for extrinsics
This is a massive simplification to the extrinsic code, and also
allows various simplifications in other parts of the code too.

This does cause one regression: wielding/unwielding Grayswandir
while hallucinating no longer prints a message. It was previously
being printed in entirely the wrong place deep in spaghetti code,
so the old implementation is really inadvisable and I'm glad it's
gone. The message hasn't been added to the correct place yet, though.
(Probably somewhere in the wield code would be the right place.)

A side effect of this code is that it's now possible for the left
and right legs to have independent woundedness timeouts. The
previous implementation stored information about which leg was
wounded in the *extrinsic* wounded legs field, which is a bad idea
for all sorts of reasons; this implementation uses the cleaner
solution of separate properties for wounded left legs and wounded
right legs.