Fix a vanilla bug and an NH4 bug: item loss while equipping
In vanilla NetHack, you can cause a segfault via equipping boots of
levitation, and getting a quantum mechanic to teleport you onto a
sink during the helplessness period. NH4 had a similar but different
way to trigger the bug, via equipping the boots while standing on the
sink; that doesn't happen in vanilla because the equip code runs in a
different order, perhaps in a partial but ultimately unsuccessful
attempt to work around this bug.

The problem backtrace is unmul() -> Boots_on(); Boots_on() assumes
that uarmg is still set from when the armor was originally equipped,
but it's possible to use the sink to NULL it out in the meantime.
This code fixes the problem in a generic way, by causing the *_on()
functions to do nothing if the item they're trying to equip has gone
missing in the meantime, or if it goes missing during their execution
(say due to spoteffects()).