Fix issues in header file dependencies (found via an aimake upgrade)
The latest version of aimake is a lot stricter about ensuring that
header files make sense. This commit removes all dependencies on
header file order within libnethack; including the header files in
any order should now work, ending the need to experiment to work out
which order they should be in (or to just cop out and use hack.h).

Interestingly, aimake caught a couple of these mistakes in libjansson
too. It's not *wrong*, per se; but it does make life harder for
maintenance programmers.