Move all extrinsic updates to go through the same function
Namely, setequip(). This is just a wrapper for the various *_on()/
*_off()/*_gone() functions for the moment. Eventually, it will
contain the shared logic for them.

This also fixes a bug in polymorph; polymorphing a worn amulet of
strangulation would not fix the strangulation.

Messages for extrinsic updates are still broken, but now at least
they're all broken in a centralized manner.