Greatly simplify the *_on/off/gone functions, fix messages
All the *_on/off/gone functions are now combined into a single
function setequip() (which was previously a wrapper around them; now
it does the work itself). setequip() is now responsible for printing
messages, meaning that they're printed correctly for all armor
equip/unequip events (regardless of whether the unequip event is
because the player requested it or for some other reason).