Fix macro expansions in libuncursed.c
Instead of expanding to prototype + K&R definition like they used to
(which is a gcc extension that makes both gcc and clang warn like
mad), they expand to a prototype + new-style definition, which is
much more portable. Instead of using semicolons to separate the
argument types, commas and an extra set of parens are used, as is
common with macros that produce declarations. (The extra set of
parens are removed with the UNCURSED_DEPAREN macro, which is written
in portable C99, and is probably easier to follow than the old code,
as well as being better-quality.)