Globals purge: stethoscopes
Stethoscope timing being a property of the dungeon is obviously
ridiculous. I had two options: make it a property of the character;
or make it a property of the stethoscope. I chose to make it a
property of the stethoscope, so that it's possible to use two
stethoscopes to do twice the HP monitoring; this does not seem
particularly broken (especially because they have to be in main
inventory to work correctly).