Manually double-buffer SDL rendering
Not for performance, but to work around an issue on Ubuntu Saucy,
where the renderer's default texture would get corrupted.

There was also a bug in the previous versions that would cause the
font to disappear when using the software renderer. I'm not longer
able to reproduce that bug, but I'm also unable to reproduce it with
previous revisions either, so it's probably still there.