Remove calls to redrawwin
Discarding assumptions about the current screen contents isn't the
sort of thing we should be doing on every window redraw. It's
already done when the player presses control-R (or whatever the
refresh command is bound to) because that includes a call to clear,
which causes assumptions to be dropped at the next doupdate(). So
we shouldn't be doing it the rest of the time. This makes updates
faster because they don't have to touch as much of the screen.