Remove level-based special casing of high altars
Now an additional bit is stored on high altars to indicate that they are, in
fact, high.

This causes an extremely subtle save breakage: all saves will work fine with
this patch, with the exception of ones on Astral, where the altars will suddenly
stop being high altars, making it imposisble to offer the Amulet.