getpos()/farlook bugfixes and new API
getpos()'s API was inconsistent; different bits of code expected it
to have different return values, meaning that extended farlook (the
"/" command) was interpreting keypresses in an off-by-one manner,
among other things. This fix adds a new, consistent API on getpos(),
and futureproofs it for when request cancelling is implemented.

These fixes also come with a new feature. Vanilla's treatement of
.,:; at the farlook and extended farlook prompts is confusing and
not as useful as it could be. There is now a consistent
interpretation:

. (and return, space etc.) does the default action for the command
  you used;
, will do a quick farlook and prompt for another cell (people often
  did this with /y in vanilla, but now it works with ; too, which is
  less awkward all round);
: will display extra information (as in vanilla);
; will display extra information and prompt for another cell.