Don't update the map or pickup items during the game creation sequence
Because the map will be updated when we restore the game immediately
after creating it.

As for the items, this is a vanilla bugfix: items on the starting
square would be picked up even with autopickup off. Now, we don't
pick them up even with autopickup on, in order to work around the
approximately 1 in 3 million chance that a player dies during the
newgame sequence due to picking up a blasting artifact (which would
leave us with no save file to load).