Let the client tell the server when to print welcomes
If there's a timeout, then the client will reconnect without sending
a welcome request, so it just looks like the situation without the
timeout. This also fixes the legacy and new game welcomes, which were
broken in a recent commit.

This has been broken ever since NitroHack (and thus probably affects
DynaHack too, although I haven't checked), but only for server play.
It's nice to finally have it fixed.